intelligent-portfolio-analyst/requirements.txt
streamlit
pandas
plotly
python-dotenv
langgraph
langchain-groq
yfinance
pdfplumber
openpyxl